requirements.txt
tensorflow-gpu==2.0.0
opencv-python
tqdm
argparse
h5py
scipy